Create common entity between Window and Capture + reorganize root modules
